+++ pages/src/navigation/header/header-search.ts
@@ -1,8 +1,7 @@
-import {prop} from '../../../../src/modules/esl-utils/decorators/prop';
+import {prop, attr, boolAttr} from '../../../../src/modules/esl-utils/decorators';
 import {CSSClassUtils} from '../../../../src/modules/esl-utils/dom/class';
 import {afterNextRender} from '../../../../src/modules/esl-utils/async/raf';
 import {parseNumber} from '../../../../src/modules/esl-utils/misc/format';
-import {attr, boolAttr} from '../../../../src/modules/esl-base-element/core';
 import {TraversingQuery} from '../../../../src/modules/esl-traversing-query/core';
 import {ESLToggleable} from '../../../../src/modules/esl-toggleable/core';
 import {requestGss} from '../../search/search-script';
