+++ pages/src/search/search.ts
@@ -1,6 +1,6 @@
 import {ESLBaseElement} from '../../../src/modules/esl-base-element/core';
 import {CSSClassUtils} from '../../../src/modules/esl-utils/dom/class';
-import {ready} from '../../../src/modules/esl-utils/decorators/ready';
+import {ready} from '../../../src/modules/esl-utils/decorators';
 import {requestGss} from './search-script';
 
 export class ESLDemoSearchPageWrapper extends ESLBaseElement {
